separated classes onto separate docs
